feat: add group roles

+++ src/components/Groups/GroupCard.tsx
@@ -1,18 +1,21 @@
-import { ApolloCache, useReactiveVar } from "@apollo/client";
+import { ApolloCache, FetchResult, useReactiveVar } from "@apollo/client";
+import { AccountBox } from "@mui/icons-material";
 import {
   Box,
   Card,
   CardContent,
   CardHeader as MuiCardHeader,
   CardProps,
+  MenuItem,
   styled,
   Typography,
 } from "@mui/material";
 import produce from "immer";
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
-import { isLoggedInVar } from "../../apollo/cache";
+import { isLoggedInVar, toastVar } from "../../apollo/cache";
 import {
+  DeleteGroupMutation,
   GroupCardFragment,
   GroupsDocument,
   GroupsQuery,
@@ -20,8 +23,11 @@ import {
 } from "../../apollo/gen";
 import {
   MIDDOT_WITH_SPACES,
+  NavigationPaths,
   TypeNames,
 } from "../../constants/common.constants";
+import { GroupPermissions } from "../../constants/role.constants";
+import { redirectTo } from "../../utils/common.utils";
 import {
   getEditGroupPath,
   getGroupMembersPath,
@@ -33,20 +39,26 @@ import Link from "../Shared/Link";
 import GroupAvatar from "./GroupAvatar";
 import JoinButton from "./JoinButton";
 
-export const removeGroup = (id: number) => (cache: ApolloCache<any>) => {
-  cache.updateQuery<GroupsQuery>({ query: GroupsDocument }, (groupsData) =>
-    produce(groupsData, (draft) => {
-      if (!draft) {
-        return;
-      }
-      const index = draft.groups.findIndex((p) => p.id === id);
-      draft.groups.splice(index, 1);
-    })
-  );
-  const cacheId = cache.identify({ id, __typename: TypeNames.Group });
-  cache.evict({ id: cacheId });
-  cache.gc();
-};
+export const removeGroup =
+  (id: number) =>
+  (cache: ApolloCache<any>, { errors }: FetchResult<DeleteGroupMutation>) => {
+    if (errors) {
+      toastVar({ status: "error", title: errors[0].message });
+      return;
+    }
+    cache.updateQuery<GroupsQuery>({ query: GroupsDocument }, (groupsData) =>
+      produce(groupsData, (draft) => {
+        if (!draft) {
+          return;
+        }
+        const index = draft.groups.findIndex((p) => p.id === id);
+        draft.groups.splice(index, 1);
+      })
+    );
+    const cacheId = cache.identify({ id, __typename: TypeNames.Group });
+    cache.evict({ id: cacheId });
+    cache.gc();
+  };
 
 const CardHeader = styled(MuiCardHeader)(() => ({
   paddingBottom: 0,
@@ -64,10 +76,15 @@ const GroupCard = ({ group, currentUserId, ...cardProps }: Props) => {
 
   const { t } = useTranslation();
 
-  const { id, name, description, members, memberRequestCount } = group;
+  const { id, name, description, members, memberRequestCount, myPermissions } =
+    group;
   const currentMember = isLoggedIn
     ? members.find(({ user }) => currentUserId === user.id)
     : undefined;
+
+  const canApproveMemberRequests = myPermissions.includes(
+    GroupPermissions.ApproveMemberRequests
+  );
 
   const editGroupPath = getEditGroupPath(name);
   const groupMembersPath = getGroupMembersPath(name);
@@ -82,26 +99,46 @@ const GroupCard = ({ group, currentUserId, ...cardProps }: Props) => {
       update: removeGroup(id),
     });
 
+  const handleRolesButtonClick = async () => {
+    const groupRolesPath = `${NavigationPaths.Groups}/${name}/roles`;
+    await redirectTo(groupRolesPath);
+  };
+
+  const renderItemMenu = () => {
+    const canDeleteGroup = myPermissions.includes(GroupPermissions.DeleteGroup);
+    const canUpdateGroup = myPermissions.includes(GroupPermissions.UpdateGroup);
+    const canManageRoles = myPermissions.includes(GroupPermissions.ManageRoles);
+    if (!canDeleteGroup && !canUpdateGroup && !canManageRoles) {
+      return null;
+    }
+
+    return (
+      <ItemMenu
+        itemId={id}
+        anchorEl={menuAnchorEl}
+        canDelete={canDeleteGroup}
+        canUpdate={canUpdateGroup}
+        deleteItem={handleDelete}
+        deletePrompt={deleteGroupPrompt}
+        editPath={editGroupPath}
+        setAnchorEl={setMenuAnchorEl}
+      >
+        {canManageRoles && (
+          <MenuItem onClick={handleRolesButtonClick}>
+            <AccountBox fontSize="small" sx={{ marginRight: 1 }} />
+            {t("roles.actions.manageRoles")}
+          </MenuItem>
+        )}
+      </ItemMenu>
+    );
+  };
+
   return (
     <Card {...cardProps}>
       <CardHeader
+        action={renderItemMenu()}
         avatar={<GroupAvatar group={group} />}
         title={<Link href={groupPath}>{name}</Link>}
-        action={
-          // TODO: Add permission logic for edit and delete
-          currentMember && (
-            <ItemMenu
-              anchorEl={menuAnchorEl}
-              deleteItem={handleDelete}
-              deletePrompt={deleteGroupPrompt}
-              editPath={editGroupPath}
-              itemId={id}
-              setAnchorEl={setMenuAnchorEl}
-              canDelete
-              canEdit
-            />
-          )
-        }
       />
       <CardContent>
         <Typography sx={{ marginBottom: 1.25 }}>{description}</Typography>
@@ -110,7 +147,8 @@ const GroupCard = ({ group, currentUserId, ...cardProps }: Props) => {
           <Link href={groupMembersPath}>
             {t("groups.labels.members", { count: members.length })}
           </Link>
-          {currentMember && (
+
+          {canApproveMemberRequests && typeof memberRequestCount === "number" && (
             <>
               {MIDDOT_WITH_SPACES}
               <Link href={memberRequestsPath}>
